Codecov: Exclude tests that rely on secrets from codecov analysis
