+++ codecov.yml
@@ -16,3 +16,12 @@ codecov:
   notify:
     wait_for_ci: false
     after_n_builds: 1
+ignore:
+  # Ignore test that requires set up of 
+  # the conda environment and is not included
+  # in the coverage report collected in the standard way
+  - "primo/demo/tests/test_demo.py"
+  # Ignore tests that rely on secrets which are 
+  # not passed to incoming PRs to avoid false 
+  # alarm on reduced coverage
+  - "primo/utils/tests/test_secrets.py"
